# public/openapi.yaml
openapi: 3.1.0
info:
  title: TrustLayer API
  description: |
    **TrustLayer** is the AI Agent Action Firewall for Enterprise Systems.

    Every consequential agent action is evaluated against identity, policy, moral alignment,
    and risk rules before execution. Ed25519-signed audit certificates provide tamper-evident
    provenance for every decision.

    ## Authentication
    All API requests require a Bearer token with prefix `tl_`:
    ```
    Authorization: Bearer tl_your_api_key_here
    ```
    Generate API keys in **Account Settings → API Keys**.

    ## Tenant Isolation
    All requests require `X-Tenant-Id` header (or use the session tenant for dashboard calls).

    ## Rate Limits
    - Starter: 1,000 req/min
    - Pro: 10,000 req/min
    - Enterprise: unlimited
  version: 2.0.0
  contact:
    name: TrustLayer Support
    url: https://trustlayer.com
servers:
  - url: https://api.trustlayer.com/api/v1
    description: Production API
  - url: http://localhost:3000/api/v1
    description: Local Development
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: tl_prod_...
...
paths:
...
  /siem/subscribe:
    get:
      summary: List SIEM Endpoints
      tags: [SIEM]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SIEM endpoint list
    post:
      summary: Register SIEM Endpoint
      description: Registers a push endpoint for real-time event forwarding.
      tags: [SIEM]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint_url]
              properties:
                name:
                  type: string
                endpoint_url:
                  type: string
                  format: uri
                format:
                  type: string
                  enum: [json, cef]
                  default: json
                auth_header:
                  type: string
                  description: "Header:Value pair, e.g. Authorization:Bearer token"
                event_filter:
                  type: array
                  items:
                    type: string
                  description: Event type filter (supports wildcards, e.g. agent.*)
      responses:
        '201':
          description: SIEM endpoint registered
    delete:
      summary: Remove SIEM Endpoint
      tags: [SIEM]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Endpoint removed
...